requirements.txt
hyperon==0.2.2
flask==3.1.2
requests==2.32.5
sqlalchemy>=1.4.0
pymysql>=1.0.0
python-dotenv>=0.19.0

# Additional production dependencies
openai>=1.0.0
africastalking>=1.2.0
gunicorn>=21.0.0
cryptography>=41.0.0